backend/app/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
psycopg2-binary
pydantic[email]==2.8.0
requests
redis>=5.0.0
python-jose[cryptography]
argon2-cffi>=23.1.0
cryptography
python-dotenv
alembic
pytest
httpx
email-validator
google-api-python-client
google-auth
google-auth-oauthlib